Moving CEED/PETSc operator construction logic into place.
