+++ include/private/rdyoperatorimpl.h
@@ -179,6 +179,16 @@ PETSC_INTERN PetscErrorCode DestroyOperator(Operator **);
 // operator timestepping function
 PETSC_INTERN PetscErrorCode ApplyOperator(Operator *, PetscReal, Vec, Vec);
 
+//--------------------------------------------------
+// CEED/PETSc Flux and Source Operator Constructors
+//--------------------------------------------------
+
+PETSC_INTERN PetscErrorCode CreateCeedFluxOperator(RDyConfig *, RDyMesh *, PetscInt, RDyBoundary *, RDyCondition *, CeedOperator *);
+PETSC_INTERN PetscErrorCode CreateCeedSourceOperator(RDyConfig *, RDyMesh *, CeedOperator *);
+PETSC_INTERN PetscErrorCode CreatePetscFluxOperator(RDyConfig *, RDyMesh *, PetscInt, RDyBoundary *, RDyCondition *, Vec *, Vec *,
+                                                    OperatorDiagnostics *, PetscOperator *);
+PETSC_INTERN PetscErrorCode CreatePetscSourceOperator(RDyConfig *, RDyMesh *, Vec, Vec, PetscOperator *);
+
 //----------------------
 // Operator Data Access
 //----------------------
